Moved commands directory up a level and rewrote require paths.
